изменила имя пакета, внесла изменения в пользовательский интерфейс
