+++ Android1_Lesson1_HomeWork/app/src/main/java/com/example/natali/android1/MainActivity.java
@@ -1,14 +1,8 @@
-package com.example.natali.android1_lesson1_homework;
+package com.example.natali.android1;
 
-import android.content.Intent;
 import android.support.v4.app.FragmentTransaction;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-import android.widget.EditText;
-
-import org.w3c.dom.Text;
 
 public class MainActivity extends AppCompatActivity {
