+++ Android1_Lesson1_HomeWork/app/src/main/java/com/example/natali/android1/WeatherFragment.java
@@ -1,4 +1,4 @@
-package com.example.natali.android1_lesson1_homework;
+package com.example.natali.android1;
 
 import android.os.Bundle;
 import android.support.annotation.NonNull;
@@ -8,6 +8,9 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import java.text.SimpleDateFormat;
+import java.util.Date;
 
 public class WeatherFragment extends Fragment {
     private static final String KEY_POSITION = "Position";
@@ -22,6 +25,12 @@ public class WeatherFragment extends Fragment {
         args.putString(KEY_NAME, name);
         weatherFragment.setArguments(args);
         return weatherFragment;
+    }
+
+    public String dataNow() {
+        Date dateNow = new Date();
+        SimpleDateFormat formatForDateNow = new SimpleDateFormat("E dd.MM.yyyy");
+        return formatForDateNow.format(dateNow);
     }
 
     public int getPosition() {
@@ -40,14 +49,17 @@ public class WeatherFragment extends Fragment {
 
         String[] weather = getResources().getStringArray(R.array.Weather);
         String[] temperature = getResources().getStringArray(R.array.Temperatures);
+        String data = dataNow();
 
         TextView cityNameView = layout.findViewById(R.id.textViewCity);
         TextView weatherView = layout.findViewById(R.id.textViewWeather);
         TextView temperatureView = layout.findViewById(R.id.textViewTemperature);
+        TextView textViewData = (TextView) layout.findViewById(R.id.textViewData);
 
         cityNameView.setText(getName());
         weatherView.setText(weather[getPosition()]);
         temperatureView.setText(temperature[getPosition()]);
+        textViewData.setText(data);
 
         return layout;
     }
