+++ Android1_Lesson1_HomeWork/app/src/main/res/layout/fragment_cities.xml
@@ -2,7 +2,9 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical">
+    android:orientation="vertical"
+    android:paddingLeft="16dp"
+    android:paddingTop="20dp">
 
     <ImageView
         android:layout_width="wrap_content"
